requirements.txt
streamlit
pdfplumber
docx2pdf
pymupdf
pillow
opencv-python
numpy
scikit-image
streamlit-drawable-canvas
